requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
tensorflow
librosa
pillow
matplotlib
opencv-python
yolo
ultralytics
